Create sortcharfreq.cpp
What did  we do here?
 Our goal is to present the new string in descending order such that ,the characters which appear more frequently are shown at first. 
For example: Tree will be outputted as eerT or eeTr, both are valid. 
So to solve this we take a map , where we map the freq and index of the characters, for this we take an ordered vector pair and store the values accordingly, next we take a max heap which also acts as our priority queue, in which we push the value which is the highest, ie the highest frequency of a string and the character itself. Later we declare an empty string and append the result into it . Till our queue is not empty,we pop out values, the one with bigger freq will get popped out first.

 Algo:
Declare Map to store the count of each char
Declare PQ to get the value stored acc to the freq
Push in  PQ from map to sort acc to the values
Traverse in heap and append char to the string and return the ans