Create MinAvgDiff.cpp
We took a vector PREFIX where we store the sum of the integers of the (n-1)th indexes  , now our work here is to partition the array and bring out the minimum differnce, so we take the index from the vector and split it and make the necesaary calculations and store the minimum value. In the end we return the index with the least value.