Create rangesumbst.cpp
Simple, easy problem , we do a base condition check if the root of the tree is null or not. After that we check we traverse the BST and sum up, how to do the checking? first we go to the left node and check if the root node (the new left node) has a value lower than the low variable, if it stays in the range of high and low , we repeat the same for the right branches and add the sum and return it .