Create leafsimilartree.cpp
Traverse from the left most node of the tree and check if null value or not. Take two vectors V1 and V2, traverse and add the last node value to the vector.Check if the two  vectors lexicographically if they are holding the same elements in same order or not. Return true or false.