Create maxprodbinarytree
Take two variables, total sum which will store the total sum of the elements in the tree and another variable , result which will store the product of the sums. We first create a recursive function to traverse the binary tree and add the sum of the elements in the total sum variable.After that we create another function to store the left and right sub treee sums by traversing.We use a logic, ie. Total sum minus sum under left subtree will give us the sum under the right subtree, this will help us to choose the correct node and the edge to be removed to maximise the product. We use a final operation in which we use the max fucntion and store the product in the result variable and finally return it as our Final Answer. Total Time Complexity is O(n) , same as Space Complexity